Sidebar theme aus sidebar.dart entfernt und in theme.dart hinzugefügt. Funktion unverändert.
