+++ lib/presentation/sidebar.dart
@@ -1,3 +1,4 @@
+import 'package:ReHome/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:sidebarx/sidebarx.dart';
 
@@ -14,45 +15,7 @@ class Sidebar extends StatelessWidget {
   Widget build(BuildContext context) {
     return SidebarX(
       controller: _controller,
-      theme: SidebarXTheme(
-        margin: const EdgeInsets.all(10),
-        decoration: BoxDecoration(
-          color: canvasColor,
-          borderRadius: BorderRadius.circular(20),
-        ),
-        hoverColor: scaffoldBackgroundColor,
-        textStyle: TextStyle(color: Colors.white.withOpacity(0.7)),
-        selectedTextStyle: const TextStyle(color: Colors.white),
-        itemTextPadding: const EdgeInsets.only(left: 30),
-        selectedItemTextPadding: const EdgeInsets.only(left: 30),
-        itemDecoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          border: Border.all(color: canvasColor),
-        ),
-        selectedItemDecoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          border: Border.all(
-            color: actionColor.withOpacity(0.37),
-          ),
-          gradient: const LinearGradient(
-            colors: [accentCanvasColor, canvasColor],
-          ),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withOpacity(0.28),
-              blurRadius: 30,
-            )
-          ],
-        ),
-        iconTheme: IconThemeData(
-          color: Colors.white.withOpacity(0.7),
-          size: 20,
-        ),
-        selectedIconTheme: const IconThemeData(
-          color: Colors.white,
-          size: 20,
-        ),
-      ),
+      theme: sidebarTheme,
       extendedTheme: const SidebarXTheme(
         // sidebar width
         width: 180,
@@ -62,22 +25,7 @@ class Sidebar extends StatelessWidget {
       ),
       footerDivider: divider,
 
-      /**
-     
-      //ReHome Logo linke untere Ecke
-      toggleButtonBuilder: (context, extended) {
-        return SizedBox(
-          height: 70,
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Image.asset('assets/ReHomeLogo.png'),
-          ),
-        );
-      },
-     */
-
       // Profilbild/Profil obere linke Ecke
-
       toggleButtonBuilder: (context, extended) {
         return const Center();
       },
@@ -100,11 +48,3 @@ class Sidebar extends StatelessWidget {
     );
   }
 }
-
-const primaryColor = Color(0xFF685BFF);
-const canvasColor = Color(0xFF2E2E48);
-const scaffoldBackgroundColor = Color(0xFF464667);
-const accentCanvasColor = Color(0xFF3E3E61);
-const white = Colors.white;
-final actionColor = const Color(0xFF5F5FA7).withOpacity(0.6);
-final divider = Divider(color: white.withOpacity(0.3), height: 1);
